test(web-mobile): extract shared reactive React harness (mp-wrp)
reset.test.jsx and admin_schedule.test.jsx each defined their own
single-component reactive React shim, with minor variations (admin
schedule added effect-cleanup tracking and a render counter). Pull
the harness into web-mobile/js/__tests__/helpers/reactive_react.js so
future component-logic tests can reuse it and the two existing
consumers stop drifting.

The unified harness is the superset: useEffect runs once and captures
cleanups, plus mount/unmount/renderCount accessors. reset.jsx's only
useEffect just returns a mount-time cleanup, so running it during
render is a no-op until unmount — behavior unchanged.
