+++ web-mobile/js/__tests__/admin_schedule.test.jsx
@@ -1,5 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { timeEdited, timeToMinutes, clampMatchDuration, filterMatchesByCourt, suggestRebalances, computeCourtPaceStats, allMatchesCompleted } from '../admin_schedule.jsx';
+import { makeReactive } from './helpers/reactive_react.js';
 
 describe('timeEdited', () => {
   // Copilot round-9 finding: AdminTWMatch.submitTime() used
@@ -499,81 +500,6 @@ describe('CourtPacePanel timer', () => {
   let realReact;
   let runtime;
   let CourtPacePanel;
-
-  function makeReactive() {
-    let hookSlots = [];
-    let hookIndex = 0;
-    let scheduledRender = null;
-    let rootProps = null;
-    let rootFactory = null;
-    let effectCleanups = [];
-    let renderCount = 0;
-
-    function rerender() {
-      hookIndex = 0;
-      renderCount++;
-      scheduledRender = rootFactory(rootProps);
-      return scheduledRender;
-    }
-
-    const reactive = {
-      createElement: (type, props, ...children) => ({ type, props, children }),
-      useState: (initial) => {
-        const i = hookIndex++;
-        if (hookSlots.length <= i) {
-          hookSlots[i] = typeof initial === 'function' ? initial() : initial;
-        }
-        const setter = (v) => {
-          hookSlots[i] = typeof v === 'function' ? v(hookSlots[i]) : v;
-          rerender();
-        };
-        return [hookSlots[i], setter];
-      },
-      useEffect: (effect, deps) => {
-        const i = hookIndex++;
-        if (hookSlots.length <= i) {
-          hookSlots[i] = deps;
-          const cleanup = effect();
-          if (typeof cleanup === 'function') {
-            effectCleanups.push(cleanup);
-          }
-        }
-      },
-      // useMemo runs eagerly without dependency tracking — the test runtime
-      // intentionally simplifies hooks for render isolation. This means tests
-      // can't catch "tick missing from deps" regressions; that contract is
-      // enforced by code review instead.
-      useMemo: (fn) => fn(),
-      useRef: (initial) => {
-        const i = hookIndex++;
-        if (hookSlots.length <= i) {
-          hookSlots[i] = { current: initial };
-        }
-        return hookSlots[i];
-      },
-      useLayoutEffect: () => {},
-      memo: (c) => c,
-    };
-
-    return {
-      React: reactive,
-      mount: (factory, props) => {
-        hookSlots = [];
-        hookIndex = 0;
-        rootFactory = factory;
-        rootProps = props;
-        effectCleanups = [];
-        renderCount = 0;
-        return rerender();
-      },
-      unmount: () => {
-        effectCleanups.forEach(c => c());
-        effectCleanups = [];
-      },
-      currentTree: () => scheduledRender,
-      renderCount: () => renderCount,
-    };
-  }
 
   beforeEach(async () => {
     realReact = global.React;
